@@ -117,7 +117,7 @@ titlecase = "^2.4.1"
 
 [tool.poetry.group.lint.dependencies]
 docstrfmt = ">=1.11.1"
-ruff = ">=0.6.4"
+ruff = ">=0.13.0"
 sphinx-lint = ">=1.0.0"
 
 [tool.poetry.group.typing.dependencies]
@@ -226,7 +226,7 @@ cmd = "make -C docs $COMMANDS"
 
 [tool.poe.tasks.format]
 help = "Format the codebase"
-cmd = "ruff format"
+cmd = "ruff format --config=pyproject.toml"
 
 [tool.poe.tasks.format-docs]
 help = "Format the documentation"
@@ -234,7 +234,7 @@ cmd = "docstrfmt docs *.rst"
 
 [tool.poe.tasks.lint]
 help = "Check the code for linting issues. Accepts ruff options."
-cmd = "ruff check"
+cmd = "ruff check --config=pyproject.toml"
 
 [tool.poe.tasks.lint-docs]
 help = "Lint the documentation"
@@ -290,10 +290,11 @@ extend-exclude = [
 ]
 
 [tool.ruff]
-target-version = "py39"
+target-version = "py310"
 line-length = 80
 
 [tool.ruff.lint]
+future-annotations = true
 select = [
     # "ARG", # flake8-unused-arguments
     # "C4", # flake8-comprehensions
@@ -305,9 +306,9 @@ select = [
     "ISC", # flake8-implicit-str-concat
     "N", # pep8-naming
     "PT", # flake8-pytest-style
-    # "RUF", # ruff
+    "RUF", # ruff
     "UP", # pyupgrade
-    "TCH", # flake8-type-checking
+    "TC", # flake8-type-checking
     "W", # pycodestyle
 ]
 ignore = [
@@ -319,6 +320,8 @@ ignore = [
 "test/plugins/test_ftintitle.py" = ["E501"]
 "test/test_util.py" = ["E501"]
 "test/ui/test_field_diff.py" = ["E501"]
+"test/util/test_id_extractors.py" = ["E501"]
+"test/**" = ["RUF001"] # we use Unicode characters in tests
 
 [tool.ruff.lint.isort]
 split-on-trailing-comma = false
