@@ -15,6 +15,7 @@
 """Tests for MusicBrainz API wrapper."""
 
 import unittest
+from typing import ClassVar
 from unittest import mock
 
 import pytest
@@ -1017,7 +1018,11 @@ class TestMusicBrainzPlugin(PluginMixin):
     plugin = "musicbrainz"
 
     mbid = "d2a6f856-b553-40a0-ac54-a321e8e2da99"
-    RECORDING = {"title": "foo", "id": "bar", "length": 42}
+    RECORDING: ClassVar[dict[str, int | str]] = {
+        "title": "foo",
+        "id": "bar",
+        "length": 42,
+    }
 
     @pytest.fixture
     def plugin_config(self):
