@@ -72,8 +72,8 @@ class RandomTest(TestHelper, unittest.TestCase):
                     print(f"{i:2d} {'*' * positions.count(i)}")
             return self._stats(positions)
 
-        mean1, stdev1, median1 = experiment("artist")
-        mean2, stdev2, median2 = experiment("track")
+        _, stdev1, median1 = experiment("artist")
+        _, stdev2, median2 = experiment("track")
         assert 0 == pytest.approx(median1, abs=1)
         assert len(self.items) // 2 == pytest.approx(median2, abs=1)
         assert stdev2 > stdev1
