@@ -19,6 +19,7 @@ import logging
 import os
 import pkgutil
 import sys
+from typing import ClassVar
 from unittest.mock import ANY, Mock, patch
 
 import pytest
@@ -46,7 +47,7 @@ from beets.util import PromptChoice, displayable_path, syspath
 
 class TestPluginRegistration(PluginTestCase):
     class RatingPlugin(plugins.BeetsPlugin):
-        item_types = {
+        item_types: ClassVar[dict[str, types.Type]] = {
             "rating": types.Float(),
             "multi_value": types.MULTI_VALUE_DSV,
         }
@@ -70,7 +71,9 @@ class TestPluginRegistration(PluginTestCase):
 
     def test_duplicate_type(self):
         class DuplicateTypePlugin(plugins.BeetsPlugin):
-            item_types = {"rating": types.INTEGER}
+            item_types: ClassVar[dict[str, types.Type]] = {
+                "rating": types.INTEGER
+            }
 
         self.register_plugin(DuplicateTypePlugin)
         with pytest.raises(
@@ -308,7 +311,9 @@ class PromptChoicesTest(TerminalImportMixin, PluginImportTestCase):
             "Enter search",
             "enter Id",
             "aBort",
-        ) + ("Foo", "baR")
+            "Foo",
+            "baR",
+        )
 
         self.importer.add_choice(Action.SKIP)
         self.importer.run()
@@ -342,7 +347,9 @@ class PromptChoicesTest(TerminalImportMixin, PluginImportTestCase):
             "Enter search",
             "enter Id",
             "aBort",
-        ) + ("Foo", "baR")
+            "Foo",
+            "baR",
+        )
 
         config["import"]["singletons"] = True
         self.importer.add_choice(Action.SKIP)
@@ -381,7 +388,8 @@ class PromptChoicesTest(TerminalImportMixin, PluginImportTestCase):
             "Enter search",
             "enter Id",
             "aBort",
-        ) + ("baZ",)
+            "baZ",
+        )
         self.importer.add_choice(Action.SKIP)
         self.importer.run()
         self.mock_input_options.assert_called_once_with(
@@ -416,7 +424,8 @@ class PromptChoicesTest(TerminalImportMixin, PluginImportTestCase):
             "Enter search",
             "enter Id",
             "aBort",
-        ) + ("Foo",)
+            "Foo",
+        )
 
         # DummyPlugin.foo() should be called once
         with patch.object(DummyPlugin, "foo", autospec=True) as mock_foo:
@@ -458,7 +467,8 @@ class PromptChoicesTest(TerminalImportMixin, PluginImportTestCase):
             "Enter search",
             "enter Id",
             "aBort",
-        ) + ("Foo",)
+            "Foo",
+        )
 
         # DummyPlugin.foo() should be called once
         with helper.control_stdin("f\n"):
